+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - Fix issue in `DefaultTimingProvider` with `Stopwatch` frequency
 - Default maximum age of an entry in the beacon cache changed from 1h 45m to 45m
 - Removed AppMon functionality
+- Maximum length of reported error/crash stacktrace has been limited to 128k.
+- Maximum length of reported error/crash reason has been limited to 1000.
+
 ## 2.2.0 [Release date: 2021-05-19]
 [GitHub Releases](https://github.com/Dynatrace/openkit-dotnet/releases/tag/v2.2.0)